requirements.txt
python-dotenv>=0.18.0
ruamel.yaml>=0.17.10
discord.py>=1.7.3
PyNaCl>=1.4.0
youtube-dl>=2021.6.6
youtube-search-python>=1.4.6
googletrans>=3.1.0a0
num2words>=0.5.10
asyncpraw>=7.3.0
speedtest-cli>=2.1.3
